+++ src/contentScript/objects/box/BoxObject.js
