+++ src/contentScript/objects/box/HoveredState.js
