+++ src/contentScript/objects/box/IdleState.js
