+++ src/contentScript/objects/box/SelectedState.js
